Move flash stats to top level of flash sorting code
